API Version now uses query parameter api-version and has a datestamp
