+++ CovidSafe/CovidSafe.API.Tests/Controllers/MessageControllers/AreaReportControllerTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-using CovidSafe.API.v1.Controllers.MessageControllers;
+using CovidSafe.API.Controllers.MessageControllers;
 using CovidSafe.DAL.Repositories;
 using CovidSafe.DAL.Services;
 using CovidSafe.Entities.Protos;
@@ -11,7 +11,7 @@ using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 
-namespace CovidSafe.API.Tests.v1.Controllers.MessageControllers
+namespace CovidSafe.API.Tests.Controllers.MessageControllers
 {
     /// <summary>
     /// Unit tests for the <see cref="AreaReportController"/> class
