+++ CovidSafe/CovidSafe.API.Tests/Controllers/MessageControllers/SeedReportControllerTests.cs
@@ -1,13 +1,12 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-using CovidSafe.API.v1.Controllers.MessageControllers;
+using CovidSafe.API.Controllers.MessageControllers;
 using CovidSafe.DAL.Repositories;
 using CovidSafe.DAL.Services;
 using CovidSafe.Entities.Protos;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
