+++ CovidSafe/CovidSafe.API/Controllers/MessageController.cs
@@ -10,15 +10,15 @@ using CovidSafe.Entities.Validation;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
-namespace CovidSafe.API.v1.Controllers
+namespace CovidSafe.API.Controllers
 {
     /// <summary>
     /// Handles <see cref="MatchMessage"/> CRUD operations
     /// </summary>
     [ApiController]
-    [ApiVersion("1", Deprecated = true)]
-    [ApiVersion("1.1")]
-    [Route("api/v{version:apiVersion}/[controller]")]
+    [ApiVersion("2020-04-14", Deprecated = true)]
+    [ApiVersion("2020-04-15")]
+    [Route("api/[controller]")]
     public class MessageController : ControllerBase
     {
         /// <summary>
@@ -56,7 +56,6 @@ namespace CovidSafe.API.v1.Controllers
         /// <response code="200">Successful request with results</response>
         /// <response code="400">Malformed or invalid request provided</response>
         /// <returns>Collection of <see cref="MatchMessage"/> objects matching request parameters</returns>
-        [ApiExplorerSettings(IgnoreApi = true)]
         [HttpPost]
         [Consumes("application/x-protobuf", "application/json")]
         [Produces("application/x-protobuf", "application/json")]
@@ -92,8 +91,10 @@ namespace CovidSafe.API.v1.Controllers
         /// </remarks>
         /// <param name="cancellationToken">Cancellation token</param>
         /// <response code="200">Successful request with results</response>
+        [ApiExplorerSettings(IgnoreApi = true)]
+        [ApiVersionNeutral]
         [HttpHead]
-        [ProducesResponseType(typeof(IEnumerable<MatchMessage>), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         [ResponseCache(NoStore = true, Location = ResponseCacheLocation.None)]
         public Task<OkResult> HeadAsync(CancellationToken cancellationToken = default)
         {
