+++ CovidSafe/CovidSafe.API/Controllers/MessageControllers/AreaReportController.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.ComponentModel.DataAnnotations;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -8,16 +9,16 @@ using CovidSafe.Entities.Validation;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
-namespace CovidSafe.API.v1.Controllers.MessageControllers
+namespace CovidSafe.API.Controllers.MessageControllers
 {
     /// <summary>
-    /// Handles requests for infected clients volunteering <see cref="BlueToothSeed"/> identifiers
+    /// Handles requests for infected clients volunteering <see cref="AreaMatch"/> identifiers
     /// </summary>
+    [ApiVersion("2020-04-14", Deprecated = true)]
+    [ApiVersion("2020-04-15")]
     [ApiController]
-    [ApiVersion("1", Deprecated = true)]
-    [ApiVersion("1.1")]
-    [Route("api/v{version:apiVersion}/Messages/[controller]")]
-    public class SeedReportController : ControllerBase
+    [Route("api/Messages/[controller]")]
+    public class AreaReportController : ControllerBase
     {
         /// <summary>
         /// <see cref="MatchMessage"/> service layer
@@ -25,39 +26,37 @@ namespace CovidSafe.API.v1.Controllers.MessageControllers
         private IMessageService _messageService;
 
         /// <summary>
-        /// Creates a new <see cref="SeedReportController"/> instance
+        /// Creates a new <see cref="AreaReportController"/> instance
         /// </summary>
         /// <param name="messageService"><see cref="MatchMessage"/> service layer</param>
-        public SeedReportController(IMessageService messageService)
+        public AreaReportController(IMessageService messageService)
         {
             // Assign local values
             this._messageService = messageService;
         }
 
         /// <summary>
-        /// Publish a <see cref="SelfReportRequest"/> for distribution among devices relevant to 
-        /// <see cref="Region"/>
+        /// Publish a <see cref="AreaMatch"/> for distribution among devices
         /// </summary>
         /// <remarks>
         /// Sample request:
         ///
-        ///     PUT /Messages/SeedReport
+        ///     PUT /Messages/AreaReport
         ///     {
-        ///         "seeds": [{
-        ///             "seed": "00000000-0000-0000-0000-000000000000",
-        ///             "sequenceStartTime": 1586406048649,
-        ///             "sequenceEndTime": 1586408048649
-        ///         }],
-        ///         "region": {
-        ///             "latitudePrefix": 74.12,
-        ///             "longitudePrefix": -39.12,
-        ///             "precision": 2
-        ///         },
-        ///         "clientTimestamp": 1586409048649
+        ///         "userMessage": "Monitor symptoms for one week.",
+        ///         "areas": [{
+        ///             "location": {
+        ///                 "latitude": 74.12345,
+        ///                 "longitude": -39.12345
+        ///             },
+        ///             "radiusMeters": 100,
+        ///             "beginTime": 1586083599,
+        ///             "endTime": 1586085189
+        ///         }]
         ///     }
         ///
         /// </remarks>
-        /// <param name="request"><see cref="SelfReportRequest"/> content</param>
+        /// <param name="request"><see cref="AreaMatch"/> to be stored</param>
         /// <param name="cancellationToken">Cancellation token (not required in API call)</param>
         /// <response code="200">Submission successful</response>
         /// <response code="400">Malformed or invalid request</response>
@@ -66,19 +65,17 @@ namespace CovidSafe.API.v1.Controllers.MessageControllers
         [Produces("application/x-protobuf", "application/json")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ResponseCache(NoStore = true, Location = ResponseCacheLocation.None)]
-        public async Task<ActionResult> PutAsync(SelfReportRequest request, CancellationToken cancellationToken = default)
+        public async Task<ActionResult> PutAsync([Required] AreaMatch request, CancellationToken cancellationToken = default)
         {
-            // Get server timestamp at request immediately
-            long serverTimestamp = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds();
-
             try
             {
-                await this._messageService.PublishAsync(request, serverTimestamp, cancellationToken);
+                // Publish area
+                await this._messageService.PublishAreaAsync(request, cancellationToken);
                 return Ok();
             }
             catch (ValidationFailedException ex)
             {
-                // Only return validation results
+                // Only return validation issues
                 return BadRequest(ex.ValidationResult);
             }
             catch (ArgumentNullException)
