+++ CovidSafe/CovidSafe.API/appsettings.json
@@ -8,6 +8,7 @@
     "MaxDataAgeToReturnDays": 14,
     "MessageContainerName": "messages_test_3"
   },
+  "DefaultApiVersion": "2020-04-15",
   "KeyVaultUrl": "https://covidsafe-api.vault.azure.net/",
   "Logging": {
     "LogLevel": {
