+++ CovidSafe/CovidSafe.Entities/Protos/Deprecated/MessageSizeResponse.cs
@@ -1,5 +1,5 @@
 ﻿#pragma warning disable CS1591, CS0612, CS3021, IDE1006
-namespace CovidSafe.Entities.Protos.Deprecated.v1
+namespace CovidSafe.Entities.Protos.Deprecated
 {
     [global::ProtoBuf.ProtoContract()]
     public partial class MessageSizeResponse : global::ProtoBuf.IExtensible
